+++ lib/ultimate_turbo_modal/helpers/view_helper.rb
@@ -6,9 +6,9 @@ module UltimateTurboModal::Helpers
       render(UltimateTurboModal.new(request:, **), &)
     end
 
-    def drawer(position: :right, overlay: true, size: UltimateTurboModal.configuration.drawer_size, **options, &block)
-      UltimateTurboModal::Base.validate_drawer_size!(size)
-      modal(drawer: position, overlay: overlay, drawer_size: size, **options, &block)
+    def drawer(position: nil, size: nil, **options, &block)
+      cfg = UltimateTurboModal.configuration.drawer_config
+      modal(_drawer_position: position || cfg.position, size: size, **options, &block)
     end
   end
 end